requirements.txt
Django==1.6
Pillow==2.6.1
PyJWT==0.3.2
argparse==1.2.1
django-comments-xtd==1.2
django-crispy-forms==1.4.0
django-phonenumber-field==0.6
django-recaptcha==1.0.2
django-registration-redux==1.1
django-thumbs==0.4
oauthlib==0.7.2
phonenumbers==7.0.1
python-openid==2.2.5
python-social-auth==0.2.1
requests==2.5.0
requests-oauthlib==0.4.2
six==1.8.0
wsgiref==0.1.2
